+++ src/main/resources/db/migration/postgresql/V7__init_user_liked_post.sql
@@ -3,7 +3,7 @@ CREATE TABLE user_liked_post
     id               UUID                        NOT NULL DEFAULT uuid_generate_v4(),
     user_id          UUID                        NOT NULL,
     post_id          UUID                        NOT NULL,
-    version          INTEGER                     NOT NULL,
+    version          BIGINT                      NOT NULL,
     created_at       TIMESTAMP WITHOUT TIME ZONE NOT NULL,
     last_modified_at TIMESTAMP WITHOUT TIME ZONE NOT NULL,
     PRIMARY KEY (id),
